Formatted code + angles in radians everywhere + Updated cmake file

+++ src/micro/Vector.h
@@ -1,23 +1,22 @@
 #ifndef VECTOR_H
 #define VECTOR_H
 
-#include <stdlib.h>
-
 ////////////////////////////////////
 ////// VECTOR IMPLEMENTATION ///////
 ////////////////////////////////////
-typedef struct {
-    void** data;
-    size_t size;
-    size_t capacity;
-    size_t dataSize;
+typedef struct
+{
+  void **data;
+  unsigned int size;
+  unsigned int capacity;
+  unsigned int dataSize;
 } Vector;
 
 extern Vector vector_create(unsigned int dataSize);
-extern void vector_free(Vector* vec);
-extern void vector_push_back(Vector* vec, void* value);
-extern void vector_pop_back(Vector* vec);
-extern void* vector_back(Vector* vec);
-extern void* vector_at(Vector* vec, unsigned int index);
+extern void vector_free(Vector *vec);
+extern void vector_push_back(Vector *vec, void *value);
+extern void vector_pop_back(Vector *vec);
+extern void *vector_back(Vector *vec);
+extern void *vector_at(Vector *vec, unsigned int index);
 
 #endif /* end of include guard: VECTOR_H */
